requirements.txt
nb-cli==0.5.3
nonebot-adapter-onebot==2.0.0a16
nonebot-plugin-apscheduler<=0.1.2
nonebot2==2.0.0a16
